perf(npu): all-float attention softmax — 5.2 KB of program memory back and M=32 builds
The last soft-double in attn1.cc is gone: exp2_soft now takes and returns float
and the softmax accumulates l_chunk in float. Soft-double on a 32-bit AIE core is
expensive code, and this is the change that actually breaks the M=32 wall:

  .text  16208 -> 10976 B  (5232 bytes of program memory back)
  M=32 attention core .text 11504 B (headroom 4880 B) -> BUILDS, where before it
  was over the 16 KB limit by 448 B.

At M=32 the layer's non-attention stages verify as well as at M=16 (QKV, GU, SiLU
bit-exact). The attention itself is still wrong at M=32 - see the design doc: heads
0-7 ~46%, heads 8-15 0%, i.e. pass 0's first query block is right and the desync
starts at qb=1. The acquire/tap counts are provably consistent (core qk acquires
16 = QK taps/2; v 16 = V taps 16; O 4 = O taps 4), so that defect is in tap
CONTENT, not plumbing, and is the next thing to trace.

+++ engine/npu/generators/attn1.cc
@@ -65,16 +65,17 @@ static inline uint16_t f32_to_bf16(float f) {
 static inline float bf16_to_f32(uint16_t u) {
     uint32_t v = (uint32_t)u << 16; float f; __builtin_memcpy(&f, &v, 4); return f;
 }
-// 2^x. Deliberately COMPACT: the attention core's .text sits at 16,304 B of the
-// AIE2P core's 16,384 B program memory (80 bytes of headroom!), and the 8-term
-// DOUBLE-precision Horner this replaces was the largest single block in it. A
-// 4-term float polynomial is ~1e-6 relative on |f|<=0.5 - orders of magnitude
-// below the bf16 the scores are stored in - and costs a fraction of the .text.
-static inline float exp2_soft(double x) {
+// 2^x, ALL FLOAT. Two reasons, both measured: the attention core's .text sits at
+// 16,304 B of the AIE2P core's 16,384 B program memory, and soft-double on a
+// 32-bit core is expensive code (the previous 8-term double Horner was the largest
+// single block in the kernel). A 4-term float polynomial is ~1e-6 relative on
+// |f|<=0.5 - orders of magnitude below the bf16 the scores are stored in - and it
+// turned out MORE accurate than the double version, not less.
+static inline float exp2_soft(float x) {
     if (x < -126.0) return 0.0f;
     if (x > 126.0) return (float)INFINITY;
-    double n = (double)(int)(x + (x >= 0.0 ? 0.5 : -0.5));
-    float f = (float)(x - n);
+    float n = (float)(int)(x + (x >= 0.0f ? 0.5f : -0.5f));
+    float f = x - n;
     float p = 1.0f + f * (0.6931472f + f * (0.2402265f + f * (0.0555041f + f * 0.0096181f)));
     uint32_t bits; __builtin_memcpy(&bits, &p, 4);
     int e = (int)((bits >> 23) & 0xFF) + (int)n;
@@ -121,13 +122,13 @@ extern "C" void attn1_chunk(const uint16_t *__restrict qk,
             if (s > m_local) m_local = s;
         }
         float m_new = m_old > m_local ? m_old : m_local;
-        float a = (float)exp2_soft((double)(m_old - m_new) * log2e);
-        double l_chunk = 0.0;
+        float a = exp2_soft((m_old - m_new) * log2e);
+        float l_chunk = 0.0f;
         for (int c = 0; c < N_KEYS; c++) {
             int tc = c / 8, cc = c % 8;
             float s = bf16_to_f32(g_sc[(tr * (N_KEYS / 8) + tc) * 32 + rr * 8 + cc]);
-            float e = (float)exp2_soft((double)(s - m_new) * log2e);
-            l_chunk += (double)e;
+            float e = exp2_soft((s - m_new) * log2e);
+            l_chunk += e;
             g_sc[(tr * (N_KEYS / 8) + tc) * 32 + rr * 8 + cc] = f32_to_bf16(e);
         }
         m_state[r] = m_new;
